+++ jettyFactory/src/main/java/io/github/factoryfx/jetty/builder/FactoryTemplateName.java
@@ -1,6 +1,9 @@
 package io.github.factoryfx.jetty.builder;
 
-//to avoid confusion with string servlet spec path
+
+/**
+ * to avoid confusion with string servlet spec path
+ */
 public class FactoryTemplateName {
     public final String name;