+++ jettyFactory/src/test/java/io/github/factoryfx/jetty/ssl/SslContextFactoryFactoryTest.java
@@ -3,7 +3,9 @@ package io.github.factoryfx.jetty.ssl;
 import io.github.factoryfx.factory.FactoryBase;
 import io.github.factoryfx.factory.SimpleFactoryBase;
 import io.github.factoryfx.factory.builder.Scope;
-import io.github.factoryfx.jetty.builder.*;
+import io.github.factoryfx.jetty.builder.FactoryTemplateName;
+import io.github.factoryfx.jetty.builder.JettyFactoryTreeBuilder;
+import io.github.factoryfx.jetty.builder.JettyServerRootFactory;
 import io.github.factoryfx.server.Microservice;
 import org.eclipse.jetty.server.Server;
 import org.junit.jupiter.api.Assertions;
@@ -15,11 +17,17 @@ import javax.ws.rs.Path;
 import javax.ws.rs.core.Response;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
 import java.net.URL;
 import java.net.URLConnection;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
 import java.security.cert.X509Certificate;
+import java.util.Properties;
 
 public class SslContextFactoryFactoryTest {
 
@@ -65,7 +73,7 @@ public class SslContextFactoryFactoryTest {
     }
 
     public static class SslContextFactoryFactoryCustom<R extends FactoryBase<?, R>> extends SslContextFactoryFactory<R> {
-        SslContextFactoryFactoryCustom(){
+        public SslContextFactoryFactoryCustom(){
             trustStore.nullable();
         }
     }
@@ -104,7 +112,7 @@ public class SslContextFactoryFactoryTest {
             URLConnection conn = url.openConnection();
             InputStream is = conn.getInputStream();
             Assertions.assertEquals("Hello World",convertStreamToString(is));
-        } catch (NoSuchAlgorithmException | KeyManagementException | IOException e) {
+        } catch (IOException e) {
             throw new RuntimeException(e);
         } finally {
             microservice.stop();
@@ -112,30 +120,34 @@ public class SslContextFactoryFactoryTest {
 
     }
 
-    public void fixUntrustCertificate() throws KeyManagementException, NoSuchAlgorithmException {
-        TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
-            public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-                return null;
-            }
+    public void fixUntrustCertificate()  {
+        try {
+            TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
+                public X509Certificate[] getAcceptedIssuers() {
+                    return null;
+                }
 
-            public void checkClientTrusted(X509Certificate[] certs, String authType) {
-            }
+                public void checkClientTrusted(X509Certificate[] certs, String authType) {
+                }
 
-            public void checkServerTrusted(X509Certificate[] certs, String authType) {
-            }
+                public void checkServerTrusted(X509Certificate[] certs, String authType) {
+                }
 
-        }};
-        SSLContext sc = SSLContext.getInstance("SSL");
-        sc.init(null, trustAllCerts, new java.security.SecureRandom());
-        HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
+            }};
+            SSLContext sc = SSLContext.getInstance("SSL");
+            sc.init(null, trustAllCerts, new java.security.SecureRandom());
+            HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
 
-        HostnameVerifier allHostsValid = new HostnameVerifier() {
-            public boolean verify(String hostname, SSLSession session) {
-                return true;
-            }
-        };
+            HostnameVerifier allHostsValid = new HostnameVerifier() {
+                public boolean verify(String hostname, SSLSession session) {
+                    return true;
+                }
+            };
 
-        HttpsURLConnection.setDefaultHostnameVerifier(allHostsValid);
+            HttpsURLConnection.setDefaultHostnameVerifier(allHostsValid);
+        } catch (NoSuchAlgorithmException | KeyManagementException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     static String convertStreamToString(java.io.InputStream is) {
@@ -143,4 +155,78 @@ public class SslContextFactoryFactoryTest {
         return s.hasNext() ? s.next() : "";
     }
 
+
+
+    //java http client only with http2+ssl?  https://stackoverflow.com/questions/52513932/force-immediate-insecure-http2-connection-with-java-httpclient
+    @Test
+    public void test_http2() {
+        JettyFactoryTreeBuilder builder = new JettyFactoryTreeBuilder((jetty,ctx)->{
+            jetty
+                    .withResource(ctx.get(TestResourceFactory.class))
+                    .setAdditionalConnector(
+                            connector-> connector.withHost("localhost").withPort(8009).withHttp2().withSsl(ctx.get(SslContextFactoryFactoryCustom.class)),new FactoryTemplateName("connector1")
+                    );
+
+        });
+        builder.addFactory(SslContextFactoryFactoryCustom.class, Scope.SINGLETON, ctx->{
+            SslContextFactoryFactoryCustom<JettyServerRootFactory> ssl = new SslContextFactoryFactoryCustom<>();
+            ssl.keyStoreType.set(KeyStoreType.jks);
+            ssl.trustStoreType.set(KeyStoreType.jks);
+            ssl.connectorType.set(SslContextFactoryFactory.ConnectorType.SERVER);
+            try (InputStream in = getClass().getResourceAsStream("/keystore.jks")){
+                byte[] bytes = in.readAllBytes();
+                ssl.keyStore.set(bytes);
+                ssl.keyStorePassword.set("password");
+
+                ssl.trustStore.set(bytes);
+                ssl.trustStorePassword.set("password");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+
+            return ssl;
+        });
+        builder.addFactory(TestResourceFactory.class, Scope.SINGLETON);
+
+        Microservice<Server, JettyServerRootFactory> microservice = builder.microservice().build();
+        try {
+            microservice.start();
+
+            {
+
+                final Properties props = System.getProperties();
+                props.setProperty("jdk.internal.httpclient.disableHostnameVerification", Boolean.TRUE.toString());
+
+                TrustManager[] trustAllCerts = new TrustManager[]{
+                        new X509TrustManager() {
+                            public java.security.cert.X509Certificate[] getAcceptedIssuers() {
+                                return null;
+                            }
+                            public void checkClientTrusted(
+                                    java.security.cert.X509Certificate[] certs, String authType) {
+                            }
+                            public void checkServerTrusted(
+                                    java.security.cert.X509Certificate[] certs, String authType) {
+                            }
+                        }
+                };
+                SSLContext sslContext = SSLContext.getInstance("TLS");
+                sslContext.init(null, trustAllCerts, new SecureRandom());
+
+//                fixUntrustCertificate();
+                HttpClient client = HttpClient.newBuilder().version(HttpClient.Version.HTTP_2).sslContext(sslContext).build();
+                HttpRequest request = HttpRequest.newBuilder().GET().uri(URI.create("https://localhost:8009/test")).build();
+                HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+                Assertions.assertEquals("Hello World", response.body());
+            }
+//
+        } catch (IOException | InterruptedException | NoSuchAlgorithmException | KeyManagementException e) {
+            throw new RuntimeException(e);
+        } finally {
+            microservice.stop();
+            final Properties props = System.getProperties();
+            props.setProperty("jdk.internal.httpclient.disableHostnameVerification", Boolean.FALSE.toString());
+        }
+
+    }
 }